+++ src/components/Admin/ManageSensors/ManageSensors.tsx
@@ -10,28 +10,22 @@ import {
   Th,
   Tr,
   Td,
-  Popover,
-  PopoverTrigger,
-  PopoverContent,
-  PopoverArrow,
-  PopoverCloseButton,
-  PopoverHeader,
-  PopoverBody,
   Center,
   Text,
-  Divider,
   HStack,
 } from '@chakra-ui/react';
 import {useAuth} from '../../../contexts/AuthContext';
 import AccessDenied from '../AccessDenied';
 import Loading from '../../Util/Loading';
 import {useTranslation} from 'react-i18next';
-import firebase, {firestore} from '../../../firebase';
+import {firestore} from '../../../firebase';
 import {DownloadCSVModal} from './DownloadData/DownloadCSVModal';
 import {AddSensorModal} from './AddSensorModal';
 import {DeleteSensorModal} from './DeleteSensorModal';
 import DeleteOldDataModal from './DeleteOldDataModal';
-import {Sensor, MoreInfoHeading} from './Util';
+import {Sensor, numberToString, timestampToDateString} from './Util';
+import {MoreInfoHeading} from './MoreInfoHeading';
+import {ToggleActiveSensorPopover} from './ToggleActivePopover';
 
 /**
  * Component for administrative page to manage the sensors.
@@ -56,10 +50,8 @@ const ManageSensors: () => JSX.Element = () => {
           querySnapshot.docs.forEach(doc => {
             if (doc.data()) {
               const sensorData = doc.data();
-              const purpleAirId: string =
-                sensorData.purpleAirId?.toString() ?? '';
               sensorList.push({
-                purpleAirId: purpleAirId,
+                purpleAirId: sensorData.purpleAirId,
                 name: sensorData.name ?? '',
                 latitude: sensorData.latitude ?? NaN,
                 longitude: sensorData.longitude ?? NaN,
@@ -78,139 +70,6 @@ const ManageSensors: () => JSX.Element = () => {
     }
     return;
   }, [isAuthenticated, isAdmin]);
-
-  /**
-   * Toggles `isActive` in a sensor's doc. This also resets the AQI buffer
-   * and PM2.5 buffer when activating or deactivating.
-   * @param event - click button event
-   * @param currentSensor - sensor for which to toggle isActive status
-   *
-   * @remarks
-   * Note that when a sensor is activated or deactivated, we do not change it in
-   * our PurpleAir group: 490. We will still get data from PurpleAir in our API
-   * call, but the resulting data will not be used anywhere.
-   */
-  function toggleActiveSensorStatus(
-    event: React.MouseEvent<HTMLButtonElement, MouseEvent>,
-    currentSensor: Sensor
-  ) {
-    event.preventDefault();
-
-    if (isAdmin) {
-      // Value from bufferStatus enum in backend
-      const bufferDoesNotExist = 2;
-
-      // Toggle the isActive and remove the buffers
-      firestore
-        .collection('sensors')
-        .doc(currentSensor.docId)
-        .update({
-          isActive: !currentSensor.isActive,
-          lastUpdated: firebase.firestore.FieldValue.serverTimestamp(),
-          aqiBufferStatus: bufferDoesNotExist,
-          aqiBuffer: firebase.firestore.FieldValue.delete(),
-          aqiBufferIndex: firebase.firestore.FieldValue.delete(),
-          pm25BufferStatus: bufferDoesNotExist,
-          pm25Buffer: firebase.firestore.FieldValue.delete(),
-          pm25BufferIndex: firebase.firestore.FieldValue.delete(),
-        })
-        .catch(() => {
-          setError(
-            t('sensors.changeActiveSensorError') + currentSensor.name ??
-              currentSensor.purpleAirId
-          );
-        });
-    }
-  }
-
-  /**
-   * Interface for ToggleActiveSensorPopover used for type safety
-   */
-  interface ToggleActiveSensorPopoverProps {
-    sensor: Sensor;
-  }
-
-  /**
-   * Creates a button that when clicked, creates a confirmation popup to change
-   * a sensor's active status. Active means that data for the sensor will be
-   * collected and shown on the map, but does not change anything in PurpleAir.
-   * @param sensor - sensor for a row
-   */
-  const ToggleActiveSensorPopover: ({
-    sensor,
-  }: ToggleActiveSensorPopoverProps) => JSX.Element = ({
-    sensor,
-  }: ToggleActiveSensorPopoverProps) => {
-    const name = sensor.name ? sensor.name : sensor.purpleAirId;
-
-    const popoverMessage =
-      (sensor.isActive
-        ? t('sensors.confirmDeactivate')
-        : t('sensors.confirmActivate')) + name;
-    const popoverNote = sensor.isActive
-      ? t('sensors.deactivateNote')
-      : t('sensors.activateNote');
-
-    return (
-      <Popover>
-        <PopoverTrigger>
-          <Button colorScheme={sensor.isActive ? 'red' : 'green'} width="full">
-            {sensor.isActive ? t('sensors.deactivate') : t('sensors.activate')}
-          </Button>
-        </PopoverTrigger>
-        <PopoverContent>
-          <PopoverArrow />
-          <PopoverCloseButton />
-          <PopoverHeader>
-            <Heading fontSize="medium">{t('common:confirm')}</Heading>
-          </PopoverHeader>
-          <PopoverBody>
-            <Text>{popoverMessage}</Text>
-            <Divider marginY={1} />
-            <Text marginBottom={1}>{popoverNote}</Text>
-            <Center>
-              <Button
-                paddingY={2}
-                colorScheme={sensor.isActive ? 'red' : 'green'}
-                onClick={event => toggleActiveSensorStatus(event, sensor)}
-              >
-                {t('common:confirm')}
-              </Button>
-            </Center>
-          </PopoverBody>
-        </PopoverContent>
-      </Popover>
-    );
-  };
-
-  /**
-   *
-   * @param timestamp - the time to convert
-   * @returns human readable date time string, unknown if null
-   */
-  function timestampToDateString(
-    timestamp: firebase.firestore.Timestamp | null
-  ) {
-    if (timestamp === null) {
-      return t('sensors.unknown');
-    } else {
-      const date: Date = timestamp.toDate();
-      return date.toLocaleDateString() + ' ' + date.toLocaleTimeString();
-    }
-  }
-
-  /**
-   *
-   * @param number - a number that can be NaN
-   * @returns human readable string for a number, 'unknown' if `NaN`
-   */
-  function numberToString(number: number) {
-    if (isNaN(number)) {
-      return t('sensors.unknown');
-    } else {
-      return String(number);
-    }
-  }
 
   if (isLoading || fetchingAuthInfo) {
     return <Loading />;
@@ -243,9 +102,11 @@ const ManageSensors: () => JSX.Element = () => {
             </HStack>
           </Center>
           <Box maxWidth="100%" overflowX="auto">
+            {/* TODO: change this to active and inactive sensors */}
             <Heading textAlign="justify" fontSize="2xl">
               {t('sensors.heading')}
             </Heading>
+            {/* TODO: make component for table of sensors */}
             <Table>
               <Thead>
                 <Tr>
@@ -279,19 +140,19 @@ const ManageSensors: () => JSX.Element = () => {
                   <Tr key={id}>
                     <Td>{sensor.name}</Td>
                     <Td>{sensor.purpleAirId}</Td>
-                    <Td>{numberToString(sensor.latitude)}</Td>
-                    <Td>{numberToString(sensor.longitude)}</Td>
+                    <Td>{numberToString(sensor.latitude, t)}</Td>
+                    <Td>{numberToString(sensor.longitude, t)}</Td>
                     <Td>
                       {sensor.isActive
                         ? t('sensors.active')
                         : t('sensors.inactive')}
                     </Td>
-                    <Td>{timestampToDateString(sensor.lastValidAqiTime)}</Td>
+                    <Td>{timestampToDateString(sensor.lastValidAqiTime, t)}</Td>
                     <Td>
-                      {timestampToDateString(sensor.lastSensorReadingTime)}
+                      {timestampToDateString(sensor.lastSensorReadingTime, t)}
                     </Td>
                     <Td>
-                      <ToggleActiveSensorPopover sensor={sensor} />
+                      <ToggleActiveSensorPopover sensor={sensor} setError={setError}/>
                     </Td>
                   </Tr>
                 ))}
